refactor: 同一レイアウトをまとめてコード削減
