+++ src/app/(one-column)/contact/page.tsx
@@ -1,0 +1,9 @@
+import ContactForm from '@/components/client/ContactForm';
+
+export default function ContactPage() {
+  return (
+    <div className="mx-auto max-w-xs sm:max-w-sm md:max-w-md">
+      <ContactForm />
+    </div>
+  );
+}
