+++ src/app/(one-column)/join/page.tsx
@@ -1,0 +1,9 @@
+import JoinForm from '@/components/client/JoinForm';
+
+export default function JoinPage() {
+  return (
+    <div className="mx-auto max-w-xs sm:max-w-sm md:max-w-md">
+      <JoinForm />
+    </div>
+  );
+}
