+++ src/app/(one-column)/layout.tsx
@@ -1,0 +1,13 @@
+import OneColumnLayout from '@/components/layout/OneColumnLayout';
+import { ReactNode } from 'react';
+
+/**
+ * 1カラムのページ共通レイアウト
+ */
+export default async function OneColumnPageLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return <OneColumnLayout>{children}</OneColumnLayout>;
+}
