+++ src/app/articles/(list)/series/[seriesId]/layout.tsx
