+++ src/app/articles/(list)/series/[seriesId]/page.tsx
