+++ src/app/news/(list)/p/[page]/page.tsx
