+++ src/components/ArticleContent.tsx
@@ -98,7 +98,7 @@ export default function ArticleContent({ article }: ArticleContentProps) {
   }`;
 
   return (
-    <section className="row-span-2 bg-white pb-12 lg:col-span-2 lg:shadow-xl">
+    <>
       {/* TODO: MathJax@3系で動かない原因を調査 */}
       <Script
         async
@@ -176,6 +176,6 @@ export default function ArticleContent({ article }: ArticleContentProps) {
         dangerouslySetInnerHTML={{ __html: article.body }}
         className="prose block w-full max-w-none px-8 text-lg leading-8 tracking-wider sm:px-16"
       ></div>
-    </section>
+    </>
   );
 }
